refactor: extract the shell builder into server/shell.js (shared with worker)
Move sliceChromeCss + shellScript + shellHtml out of server.js into a
dependency-free server/shell.js. All environment-specific inputs (file
contents, identity, config) are injected by the caller, so the same pure
builder renders the shell in the local server (which require()s it) and,
next, in the Cloudflare worker (which inlines it as code — Workers ban
eval, so shell.js also publishes itself on globalThis.TDOC_SHELL_BUILDER).

This is the prerequisite for Step 7 worker parity: one source of truth for
the shell document + client script, no drift between local and production.

Pure refactor — server.js's shellDocument is now a thin wrapper that reads
files/identity and calls SHELL.shellHtml. 20/20 boundary, 43/43 offline
unchanged.
